decimal to double

+++ LifelogBb/ApiDTOs/StrengthTrainings/StrengthTrainingOutput.cs
@@ -10,7 +10,7 @@ namespace LifelogBb.ApiDTOs.StrengthTrainings
 
         public int Reps { get; set; }
 
-        public decimal Weight { get; set; }
+        public double Weight { get; set; }
 
         public string? Notes { get; set; }
